requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.14
# by the following command:
#
#    pip-compile requirements.in
#
beautifulsoup4==4.14.3
    # via -r requirements.in
certifi==2026.1.4
    # via requests
charset-normalizer==3.4.4
    # via requests
idna==3.11
    # via requests
lxml==6.0.2
    # via -r requirements.in
python-dotenv==1.2.1
    # via -r requirements.in
requests==2.32.5
    # via -r requirements.in
soupsieve==2.8.2
    # via beautifulsoup4
typing-extensions==4.15.0
    # via beautifulsoup4
urllib3==2.6.3
    # via requests
